# move uniform_ from TH to ATen and optimize with MKL VSL