--- a/aten/src/ATen/native/native_functions.yaml
+++ b/aten/src/ATen/native/native_functions.yaml
@@ -4363,7 +4363,7 @@
 - func: uniform_(Tensor(a!) self, float from=0, float to=1, *, Generator? generator=None) -> Tensor(a!)
   variants: method
   dispatch:
-    CPU: legacy::cpu::_th_uniform_
+    CPU: uniform_cpu_
     CUDA: uniform_cuda_
   supports_named_tensor: True