This PR moves `uniform_` CPU implementation from TH to ATen, e.g. `uniform_cpu_`.
When PyTorch is built with MKL support and CPU is deteced to be intel CPU, `uniform_cpu_` will finally call MKL VSL random generator which assures best performance on intel CPUs.
Otherwise (MKL is not built or AMD CPU), `uniform_cpu_` will end up in sequential random generator from ATen, this path is identical to old TH implementation.

Single socket run speedup: **2.4x~142x**
Single core run speedup: **3.2x~13x**