@mingfeima 2019-12-09
### performance
Below shows the performance compare between original and this pr, the benchmark will compare from:

1. a range of input size: [1000], [10000], [100000], [1000000]
2. single socket and single core

benchmark code available at this [op_bench-py](https://github.com/mingfeima/op_bench-py), to reproduce, `./run.sh uniform.py`. Test machine is Xeon Skylake 6148 with 2*20 cores @2.40GHz

benchmark uses the metric of **numbers processed per second**, aka the throughput, the higher the better.

#### single socket

input size | original | this pr | speedup
-- | -- | -- | --
[1000] | 0.157 | 0.381 | 2.43
[10000] | 0.188 | 2.093 | 11.13
[100000] | 0.191 | 11.545 | 60.45
[1000000] | 0.191 | 27.207 | 142.45

#### single core

input size | original | this pr | speedup
-- | -- | -- | --
[1000] | 0.160 | 0.508 | 3.18
[10000] | 0.188 | 1.802 | 9.59
[100000] | 0.191 | 2.436 | 12.75
[1000000] | 0.191 | 2.517 | 13.18

### Validation
To prove the MKL VSL random generator as functional as ATen implementation, test cases are also provided in `uniform.py`.

The test case uses [T-test](http://geog.uoregon.edu/GeogR/topics/ttest.pdf) to verify the **mean** and **variance** of MKL VSL output aligns with theoretical **mean** and **variance** for [uniform distribution](https://en.wikipedia.org/wiki/Uniform_distribution_(continuous)):
```
for X~U(a, b)
E(X) = (a+b)/2
VAR(X) = E(X^2) - E(X)^2 = (b-a)^2 / 12
```

I do cross check the test cases with ATen implementation and MKL VSL implementation, both passed.