@XiaobingSuper 2019-12-11
#24781 Migrate uniform_ from the TH to Aten (CPU)