@VitalyFedyunin 2020-02-11
I had no time to check in details, but errors looks like related to the introduced change and I cannot merge it as is.