@mingfeima 2020-02-13
> I had no time to check in details, but errors looks like related to the introduced change and I cannot merge it as is.

yes, i was aware of the errors are real... Somehow I have some trouble in understanding the test cases. Still WIP.