@mingfeima 2020-02-25
Remove `vsl` and use plain CPU implementation for now. Will put `vsl` part in following work.